Updates for v3 of the HIBP API and one new feature.
The HaveIBeenPwned API now requires an API key.  Details on that can be found at https://www.troyhunt.com/authentication-and-the-have-i-been-pwned-api . The API key is provided via a text file that the user needs to create and supply via the new -a switch.

The new -n switch will return only the name of a breach.  If a user does not supply the -n switch, then the complete breach data will be returned.

Added checks for 401 and 503 HTTP errors.

Updated the README.md to reflect the necessity of supplying an API key via a text file.