A little bit of code house-keeping.
For greater readibility, I added a new function called printHTTPErrorOutput().  It reduced the amount of if-statements in get_results() to one if/else-if statement.  This function will also terminate the script if an invalid API key was used.

Replaced the 'magic number' in the sleep() statement with a constant, that can be conveniently found at the top of this script.