Updated to be compatible with Python 3
checkpwnedemails.py can now run on both Python 2 and Python 3.
In order to make it compatible with Python 3, all the print statements have been replaced with print() statements.
Also, the urllib and urllib2 libraries have been replaced with the requests library.
Reger to the 'Dependencies' section of README.md for more details.