Bug fix for single email lookup
The variable, named emails, was not actually being converted to a tuple with one element.  This would pass emails as a string, instead of the intended tuple, to get_results().  Thus, making get_results() iterate through each character in emails.  This was due to  misuse of the tuple() constructor.  tuple() requires a list as an argume
nt, not a string.