Refactored and with less spelling mistakes
Corrected spelling mistakes in some of the comments.
Reading of the API key from text file has been moved to it's own function.  This was done to improve modularity and readibility of the code.
Moved the printing of results (to standard output), from the get_results() function to it's own function.  Functions should do only one thing.